+++ CHANGELOG.rst
@@ -44,7 +44,7 @@ Containers
 - Starting and restarting a container now removes old container and creates a new one (#72)
 - Detail page now allows for managing the container (#72)
 - Extended container logs with stack trace in case of unknown error
-
+- Added ``ContainerActionLock`` model to throttle actions performed on a container (#75)
 
 v0.1.3 (2021-06-09)
 ===================